feat: add basic core cli subcommands and tests
