--- internal/cli/commands_test.go
@@ -1,1 +1,0 @@
-package cli
