+++ justfile
@@ -16,3 +16,6 @@ bench:
 
 lint:
     golangci-lint run ./...
+
+format:
+    golangci-lint fmt ./...
